+++ app/templates/settings/labs.hbs
@@ -335,19 +335,6 @@
                         </div>
                     </div>
                 </div>
-                <div class="gh-expandable-block">
-                    <div class="gh-expandable-header">
-                        <div>
-                            <h4 class="gh-expandable-title">Snippet replacements</h4>
-                            <p class="gh-expandable-description">
-                               When creating a snippet, allow for replacing an existing snippet.
-                            </p>
-                        </div>
-                        <div class="for-switch">
-                           <GhFeatureFlag @flag="snippetReplacements" />
-                        </div>
-                    </div>
-                </div>
             </div>
         </div>
         {{/if}}
