+++ lib/koenig-editor/addon/components/koenig-snippet-input.hbs
@@ -1,10 +1,25 @@
 <div class="kg-input-bar absolute z-999" style={{this.style}} {{did-insert this.registerAndPositionElement}} ...attributes>
-    <input
-        placeholder="Snippet name"
-        value={{this.name}}
-        class="kg-link-input pa2 pr6 mih-100 ba br3 shadow-2 f8 lh-heading tracked-2 outline-0 b--green h10 nudge-top--8"
-        {{on "input" this.nameInput}}
-        {{on "keydown" this.nameKeydown}}
-        {{did-insert this.focusInput}}
-    />
+    <GhInputWithSelect
+        @triggerClass="kg-link-input pa2 pr6 mih-100 ba br3 shadow-2 f8 lh-heading tracked-2 outline-0 h10 nudge-top--8 vertical"
+        @dropdownClass="gh-snippet-dropdown"
+        @placeholder="Snippet name"
+        @options={{@snippets}}
+        @showCreate={{true}}
+        @searchEnabled={{false}}
+        @searchField="name"
+        @onChange={{this.selectSnippet}}
+        @renderInPlace={{false}}
+        @autofocus={{true}}
+        @onInput={{this.nameInput}}
+        as |snippet|
+    >
+        {{#if snippet.__isSuggestion__}}
+            {{snippet}}
+        {{else}}
+            <div class="flex justify-between items-center">
+                {{snippet.name}}
+                <span class="fill-darkgrey dib w4 h4">{{svg-jar "sync"}}</span>
+            </div>
+        {{/if}}
+    </GhInputWithSelect>
 </div>
