--- lib/koenig-editor/app/components/koenig-snippet-input-labs.js
@@ -1,1 +1,0 @@
-export {default} from 'koenig-editor/components/koenig-snippet-input-labs';
